feat: LocateAnything-3B をサブプロセスワーカー方式に変更
remote code が transformers v5 と非互換 (rotary バッファ消失や
フォワード経路の崩壊で出力が壊れる) のため、モデルカード指定の
transformers==4.57.1 を vendor/ に隔離して常駐ワーカーで推論する。
赤ボール合成画像での検出テストに合格。

+++ mosaic_editor/detect/sam3_concept.py
@@ -71,7 +71,7 @@ class Sam3ConceptSegmenter:
                     progress_cb(f"SAM3 検出中 [{done}/{n_total}] 「{prompt}」...")
                 inputs = self.processor(
                     images=image, text=prompt, return_tensors="pt"
-                ).to(self.device)
+                ).to(self.device, dtype=self.dtype)
                 with torch.no_grad():
                     outputs = self.model(**inputs)
                 results = self.processor.post_process_instance_segmentation(
